Provide an advanced shrink-wrapping interface
In order to extend shrink-wrapping we need to make some changes to the
interface.

The goal here would be to provide an interface that would allow us to
develop shrink-wrapping extensions in-tree, and being able to switch
between implementations without having to handle both interfaces in PEI.

In order to be able to simply re-use this, we need to know which basic
blocks **use* which element (as an input), and which basic blocks are
the best suited for saving / restoring each element (as an output).

For that, the interface is split in two parts:

* ShrinkWrapInfo: it is an interface that determines the "uses" to be
used by the shrink-wrapping algorithm.
* ShrinkWrapper: it is an interface for the shrink-wrapping algorithm.
It makes it easy to switch between the current one (dominator-based) and
other ones like Fred Chow's.

Differential Revision: https://reviews.llvm.org/D41765