[CodeGen] Use RegMaskNames or CustomRegMask when printing regmasks in -debug output
The MIRPrinter prints RegMask names when possible, and if not, it prints
custom reg masks using the following syntax:

CustomRegMask(%r1,%r2,%r3).

In MO.print, use this syntax instead of:

<regmask %r1 %r2 %r3>

Also, this adds some code to search for the RegMask name and use it if
the operand is not printed in standalone mode.